travel_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
httpx==0.24.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.0.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
requests==2.31.0
ffmpeg-python==0.2.0
openai
python-multipart
websocket-client

